separated plugin functions and helpers
