+++ src/components/popup/popup.vue
@@ -132,8 +132,8 @@ import { Settings } from '../../interfaces/StorageSettings'
 import { onMounted, ref } from 'vue'
 import Switcher from './switcher.vue'
 
-import getSettings from '../../functions/getSettings'
-import wakeUpBackground from '../../functions/wakeUpBackground'
+import getSettings from '../../helpers/getSettings'
+import wakeUpBackground from '../../helpers/wakeUpBackground'
 
 const settings = ref(null as Settings)
 const version = ref('')
