+++ src/helpers/getSettings.ts
