+++ src/helpers/wakeUpBackground.ts
